List available collections

List available collections

Usage

list_collections()

Examples

list_collections()
#> $`noaa-monthly-max-dhw`
#> ###Collection
#> - id: noaa-monthly-max-dhw
#> - title: NOAA Degree Heating Week (DHW) - Monthly Aggregation
#> - description: 
#> The NOAA Coral Reef Watch (CRW) daily global 5km satellite coral bleaching Degree Heating Week (DHW) product shows accumulated heat stress, which can lead to coral bleaching and death. The scale ranges from 0 to 20 °C-weeks. The DHW product accumulates the instantaneous bleaching heat stress, measured by CRW's Coral Bleaching HotSpot, during the most recent 12-week period. It is directly related to the timing and intensity of coral bleaching.
#> - field(s): 
#> type, id, stac_version, description, links, stac_extensions, title, extent, license, keywords, providers, summaries
#>